Added MakeVars File
The solution to the Fortran problem was that I needed to add a MakeVars
file in the src folder.  Thanks to Roger again for allowing us to
include those 2 Fortran files (I didn't need to include any more).

Now to go through and make sure that all the documentation lines up.  I
got 2 Warnings and 1 Note when I compiled.  I am trying to fix these
issues.