Edit the help file for the function QICD
Originally, the help file said the beta_initial was supposed to be an
initial estimate for the entire coefficient vector (intercept included).
This is not the case.  The intercept shold not be included, and I have
made the edit.

I'm using rqPen for a variety of things and it took me a while to
realize that I wrote the help file incorrectly.