QICD edits
Made some edits to Adamsfunc.
Added a file called QICD_functions.R.  This contains all the QICD
functions.  Adamsfunc can be ignored; I left it here just in case we
want to refer to it.  We need to check to make sure that all inputs and
outputs from the QICD functions are appropriate.

Rewrote the cpp file.  It contains a better version of the QICD function
and it also contains a C++ version of the derivatives of the penalty
functions.  It is much faster and makes a noticeable difference for
large p.

We should skype to make sure we can "glue" everything together properly.